+++ mrerr/error_caller.go
@@ -1,0 +1,14 @@
+package mrerr
+
+var (
+	caller = NewCaller(
+		CallerOptions{
+			Deep:         2,
+			UseShortPath: true,
+		})
+)
+
+// SetCallerOptions - WARNING: use only when starting the main process
+func SetCallerOptions(opt CallerOptions) {
+	caller = NewCaller(opt)
+}